requirements.txt
pdfplumber>=0.9.0
pytesseract>=0.3.10
Pillow>=9.0.0
python-docx>=0.8.11